Fixing disjoint animation and refactoring.
Timer suspension when Gas is attacking seems to be working for me. I no longer observe this disjoint animation (axe staying in place). Also, everything related to Gas is brough to his own script, since souls_enemy is the generic class for all the enemies and checking if it is 'gascoigne' all the time does not seem right..